+++ vpype/__init__.py
@@ -1,6 +1,5 @@
 from .config import *
 from .decorators import *
-from .filters import *
 from .geometry import *
 from .io import *
 from .layers import *
